Split up network.c
The file was becoming pretty long. This commit moves some functions to
cmd.c and hostname.c.
